Updated to double and added tests for item selection.

+++ app/src/main/java/com/revolut/rx/RxExtensions.kt
@@ -7,7 +7,7 @@ import com.google.gson.internal.LinkedTreeMap
 import io.reactivex.Observable
 import io.reactivex.schedulers.Schedulers
 
-typealias RatesMap = LinkedTreeMap<String, Float>
+typealias RatesMap = LinkedTreeMap<String, Double>
 
 fun <T> Observable<T>.uiSubscribe(schedulers: SchedulersProvider): Observable<T> {
     return subscribeOn(schedulers.background).observeOn(schedulers.main)
